+++ src/calc_dists.cpp
@@ -4,7 +4,7 @@
 #include "utils.hpp"
 #include <highfive/H5File.hpp>
 
-using matrix_t = Eigen::MatrixXd;
+using matrix_t = Eigen::MatrixXf;
 
 template <typename VC>
 matrix_t calculate_cluster_distance(parser::cli_arguments arguments, const size_t nr_cores){
